added music Detected

+++ code/Client/src/client/HttpConnector_win.cpp
@@ -160,7 +160,7 @@ void HttpConnection::call(HttpAction op, const string& param, const string& quer
 	}
 	string header = headerStream.str();
     int sendResult = send(conn, header.c_str(), header.length(), 0);
-    if (sendResult != header.length())
+    if (sendResult != (int)header.length())
     	cout << "send returns other than " << header.length() << " bytes:" << sendResult << endl;
 
     int contentLength = -1;
